Remove old buildviz logic
With the new streaming mechanism we cannot afford sending events twice, so let's compromise on the "safe" resuming behaviour by replying jobs around the last cut-off date.
Lets for now assume that if stopped the sync at time X all the builds for that time were synced in the last sync, and no new builds happened to slip in on the very second.

Also, as we now explicitly write the last sync time, we don't handle cancellation gracefully. That needs a different explicit mechanism.