Initial support for syncing triggered-by info for Concourse
This probably needs some documentation. Basic idea:
From the job config find all the resources which trigger the build which are bound to a passing build of another job.
Find all builds connected to the same resource's version (via the resources API). One of the builds of the other job in there must have triggered our build.